feat: rename use case package
